+++ packages/deploy-configurations/types/deployment-config/tokens.ts
@@ -78,3 +78,4 @@ export type Tokens =
   | 'YIELDETH'
   | 'YFI'
   | 'ZRX'
+  | 'PYUSD'
