+++ packages/dma-library/src/strategies/common/erc4626/withdraw.ts
@@ -1,37 +1,16 @@
 import { ADDRESSES, SystemKeys } from '@deploy-configurations/addresses'
-import { Address } from '@dma-common/types'
 import { amountToWei } from '@dma-common/utils/common'
 import { operations } from '@dma-library/operations'
 import { getGenericSwapData } from '@dma-library/strategies/common'
-import { SummerStrategy } from '@dma-library/types'
+import {
+  Erc4626CommonDependencies,
+  Erc4626WithdrawPayload,
+  Erc4626WithdrawStrategy,
+} from '@dma-library/types'
 import { encodeOperation } from '@dma-library/utils/operation'
 import { isCorrelatedPosition } from '@dma-library/utils/swap'
 import { views } from '@dma-library/views'
-import { Erc4646ViewDependencies } from '@dma-library/views/common/erc4626'
-import { Erc4626Position } from '@dma-library/views/common/types'
 import BigNumber from 'bignumber.js'
-
-import { Erc4626CommonDependencies } from './deposit'
-
-export interface Erc4626WithdrawPayload {
-  returnTokenSymbol: string
-  returnTokenPrecision: number
-  returnTokenAddress: Address
-  withdrawTokenSymbol: string
-  withdrawTokenPrecision: number
-  withdrawTokenAddress: Address
-  amount: BigNumber
-  vault: string
-  proxyAddress: Address
-  user: Address
-  slippage: BigNumber
-  quoteTokenPrice: BigNumber
-}
-
-export type Erc4626WithdrawStrategy = (
-  args: Erc4626WithdrawPayload,
-  dependencies: Erc4626CommonDependencies & Erc4646ViewDependencies,
-) => Promise<SummerStrategy<Erc4626Position>>
 
 export const withdraw: Erc4626WithdrawStrategy = async (args, dependencies) => {
   const addresses = { tokens: { ...ADDRESSES[dependencies.network][SystemKeys.COMMON] } }
