+++ packages/dma-library/src/types/index.ts
@@ -124,6 +124,14 @@ export { MorphoBluePosition }
 export type { MorphoBlueMarket }
 
 export type {
+  Erc4626CommonDependencies,
+  Erc4626DepositPayload,
+  Erc4626DepositStrategy,
+  Erc4626StrategyAddresses,
+  Erc4626WithdrawPayload,
+  Erc4626WithdrawStrategy,
+} from './common'
+export type {
   EarnCumulativesData,
   EarnCumulativesRawData,
   LendingCumulativesData,
